api/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
coverage
flake8
gunicorn
django-cors-headers
django-extensions
django-filter